# specs/002_api_specification.yaml
...
openapi: "3.1.0"
info:
  title: Cyber Chakra Forensics Admin Portal API
  version: "1.0.0"
  description: |
    REST API for the CMF Admin Portal and the public-facing desktop app endpoints.
    All admin endpoints require Bearer token authentication.
    All public endpoints require a valid license key (X-License-Key header).
servers:
  - url: https://cyberchakra.online
    description: Production
  - url: https://staging.cyberchakra.online
    description: Staging

# =============================================================================
# Security Schemes
# =============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin portal JWT token obtained from /api/v1/auth/login.

    LicenseKeyAuth:
      type: apiKey
      in: header
      name: X-License-Key
      description: License key for public desktop app endpoints.

  # ===========================================================================
  # Shared Schemas
  # ===========================================================================
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 25
        total:
          type: integer
          example: 142
        total_pages:
          type: integer
          example: 6
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
  parameters:
...
    QueryPage:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    QueryPerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/downloads:
    get:
      tags: [Downloads]
      summary: List download records
      parameters:
        - $ref: '#/components/parameters/QueryPage'
        - $ref: '#/components/parameters/QueryPerPage'
        - name: release_id
          in: query
          schema:
            type: string
            format: uuid
        - name: organization_id
          in: query
          schema:
            type: string
            format: uuid
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Download records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      description: "Download record with release info"
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /api/v1/downloads/stats:
    get:
      tags: [Downloads]
      summary: Download statistics
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: group_by
          in: query
          schema:
            type: string
            enum: [day, week, month, version, platform]
            default: day
      responses:
        "200":
          description: Download stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_downloads:
                    type: integer
                  completed_downloads:
                    type: integer
                  unique_organizations:
                    type: integer
                  breakdown:
                    type: array
                    items:
                      type: object

  # ===========================================================================
  # SUPPORT TICKETS
  # ===========================================================================
...
  /api/public/v1/download/{releaseId}/{platform}:
    get:
      tags: [Public - Downloads]
      summary: Download a release binary
      security: []
      description: |
        Serves the installer/binary for the requested release and platform.
        Tracks the download. Returns a redirect to the actual file URL (S3/R2
        presigned URL) or streams the file directly for local storage.
      parameters:
        - name: releaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: platform
          in: path
          required: true
          schema:
            type: string
            enum: [windows_x64, windows_arm64, linux_x64, linux_arm64, macos_x64, macos_arm64]
        - name: license_key
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to download URL
          headers:
            Location:
              schema:
                type: string
                format: uri
            X-File-SHA256:
              schema:
                type: string
            X-File-Size:
              schema:
                type: integer
        "404":
          description: Release or platform not found
...